"feat : 코드 리팩토링  1. auth.py로 로그인 및 사용자 검증 로직 미들웨어로 분리  2. data_validator.py로 패턴 확인 필요 모듈 분리  3. database.py로 추후 실제 DB 연동 구현 용이화  4. users.py로 APIrouter 기능 활용하여 구현  5. main.py는 수정하여 실제 FastAPI 서버 구동만 되도록 수정"
