+++ main.py
@@ -1,319 +1,1 @@
-import re
-from typing import Annotated
-from fastapi import FastAPI, HTTPException, Form, UploadFile, File, Depends
-from pydantic import EmailStr
-from passlib.context import CryptContext
-from datetime import datetime, timedelta, timezone
-from jose import jwt, JWTError, ExpiredSignatureError
-import os
-from dotenv import load_dotenv
-from fastapi.security import OAuth2PasswordBearer
-
-app = FastAPI()
-#----- 데모용 DB (현재는 임시 리스트) ----------------------------
-demo_db = []
-
-#----- 닉네임 생성  -----------------------------------------
-nickname_pattern = "^[a-zA-Z0-9가-힣].{5,10}$"
-def validate_nickname(nickname: str):
-    if not re.match(nickname_pattern, nickname):
-        raise HTTPException(
-            status_code=422,
-            detail="닉네임은 5~10자 길이를 준수해야하며, 숫자, 영문 소문자, 영문 대문자, 한글을 사용할 수 있습니다. 중복 허용 불가입니다."
-        )
-
-#----- user_id 부여 (추후 DB 확장성 고려, 함수 형태로 구현)-------
-user_counter = 0
-def generate_user_id():
-    global user_counter
-    user_counter += 1
-    return f"user_{user_counter}"
-
-#----- 이미지 파일 용량 확인 ----------------------------------
-ALLOWED_TYPES = {"image/jpg", "image/png", "image/tiff", "image/bmp", "image/gif"}
-MAX_FILE_SIZE = 10 * 1024 * 1024 # 10MB
-async def validate_image_size(upload_file):
-    size = 0
-    chunk_size = 1024 * 1024  # 1MB
-    while True:
-        chunk = await upload_file.read(chunk_size)
-        if not chunk:
-            break
-        size += len(chunk)
-        if size > MAX_FILE_SIZE:
-            raise HTTPException(
-                status_code=413,
-                detail="이미지 파일은 10MB를 넘을 수 없습니다. "
-            )
-    await upload_file.seek(0)
-
-#----- 비밀번호 규칙 -------------------------------------------
-password_pattern = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[!@#$%^&*+=-]).{5,20}$"
-def validate_password(password: str):
-    if not re.match(password_pattern, password):
-        raise HTTPException(
-            status_code=422,
-            detail="비밀번호는 5~20자 길이를 준수해야하며, 숫자, 영문 소문자, 영문 대문자, 특수 문자가 각각 최소 1개씩 포함되어야 합니다."
-        )
-
-#----- 비밀 번호 해싱 from JEFF --------------------------------
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-# 1. 비밀번호 해싱
-def hash_password(password: str) -> str:
-    return pwd_context.hash(password)
-# 2. 비밀번호 검증
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    return pwd_context.verify(plain_password, hashed_password)
-
-#----- 로그인 횟수 검증 -----------------------------------------
-login_attempts = {}
-MAX_ATTEMPTS = 5 # 5회 이상 시도 시 429 에러 반환
-
-#----- 액세스 토큰 생성(로그인 메서드) ----------------------------
-load_dotenv() # 환경변수 파일 불러옴
-def create_access_token(data: dict, expires_delta: timedelta | None = None) -> str:
-    to_encode = data.copy()
-    expire = datetime.now(timezone.utc) + (expires_delta or timedelta(minutes=int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES"))))
-    to_encode.update({"exp": expire})
-    return jwt.encode(to_encode, os.getenv("SECRET_KEY"), algorithm=os.getenv("ALGORITHM"))
-
-#----- DB에서 user_id를 활용하여 사용자 찾기(코드 리뷰 반영) ---------
-def find_user_by_id(user_id: str) -> dict | None:
-    for user in demo_db:
-        if user["user_id"] == user_id:
-            return user
-    return None
-
-#----- DB에서 email_address를 활용하여 사용자 찾기(코드 리뷰 반영) ---------
-def find_user_by_email(email: str) -> dict | None:
-    """demo_db에서 email로 사용자를 조회합니다."""
-    for user in demo_db:
-        if user["email_address"] == email:
-            return user
-    return None
-
-#----- Oauth 2.0 Bearer 선언 & 토큰 디코딩 및 로그인 인증 --------- TODO : 추후 인증 미들웨어로 따로 분리 해야함!!
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/token")
-def auth_user_with_token_decoding(token: str = Depends(oauth2_scheme)):
-
-    try:
-        payload = jwt.decode(token, os.getenv("SECRET_KEY"), algorithms=[os.getenv("ALGORITHM")])
-
-    except ExpiredSignatureError:
-        raise HTTPException(status_code=401, detail="토큰이 만료되었습니다.")
-
-    except JWTError:
-        raise HTTPException(status_code=401, detail="토큰 인증 정보가 유효하지 않습니다. ")
-
-    user_id: str = payload.get("sub")
-    token_type = payload.get("type", "access")
-
-    if token_type != "access":
-        raise HTTPException(status_code=401, detail="유효한 액세스 토큰이 아닙니다.")
-
-    if user_id is None:
-        raise HTTPException(status_code=401, detail="토큰 인증 정보가 유효하지 않습니다. ")
-
-    user = find_user_by_id(user_id)
-    if user is None:
-        raise HTTPException(status_code=401, detail="일치하는 회원 정보가 없습니다. ")
-    return user
-
-
-# ----- 첫번째. 회원 가입 메서드 시작
-@app.post("/users")
-async def post_users(
-        email_address: Annotated[EmailStr, Form()],
-        password: Annotated[str, Form()],
-        nickname: Annotated[str, Form()],
-        profile_image: Annotated[UploadFile | None, File()] = None
-):
-    # 이메일 중복 확인
-    if any(user["email_address"] == email_address for user in demo_db):
-        raise HTTPException(status_code=409, detail="해당 이메일은 이미 등록되어 있습니다.")
-    # 비밀번호 규칙
-    validate_password(password)
-    # 비밀번호 해싱
-    hashed_password = hash_password(password)
-    # 닉네임 규칙
-    validate_nickname(nickname)
-    # 닉네임 중복 확인
-    if any(user["nickname"] == nickname for user in demo_db):
-        raise HTTPException(status_code=409, detail="해당 닉네임은 이미 등록되어 있습니다.")
-    # 코드 리뷰 반영 --> 프로필 이미지 처리 로직 추가
-    profile_image_url = None
-    # 이미지 파일 확장자 검증
-    if profile_image:
-        if profile_image.content_type not in ALLOWED_TYPES:
-            raise HTTPException(415, "지원하지 않는 이미지 파일 형식입니다. ")
-        #TODO : 파일 저장 및 이미지 크기 확인 로직
-        await validate_image_size(profile_image)
-        profile_image_url = f"/uploads/{profile_image.filename}"
-    # user_id 부여
-    user_id = generate_user_id()
-    # DB에 저장 -- 지금은 리스트에 임시로...
-    demo_db.append({
-        "user_id": user_id,
-        "email_address": email_address,
-        "hashed_password": hashed_password,
-        "nickname": nickname,
-        "profile_image_url": profile_image_url,
-        "users_created_time" : datetime.now(timezone.utc).strftime('%Y.%m.%d - %H:%M:%S'),
-        "users_modified_time" : None
-    })
-    return {
-        "status": "success",
-        "message": "회원가입이 정상적으로 처리되었습니다.",
-        "data": {
-            "user_id": user_id,
-            "email_address": email_address,
-            "nickname": nickname,
-            "profile_image_url": profile_image_url,
-            "users_created_time": datetime.now(timezone.utc).strftime('%Y.%m.%d - %H:%M:%S')
-        }
-    }
-
-# ----- 두번째. 로그인 메서드 시작
-@app.post("/auth/token")
-async def post_auth_token(
-        email_address: Annotated[EmailStr, Form()],
-        password: Annotated[str, Form()],
-):
-    # 1. 이메일 가입 여부 확인
-    is_users_email = None
-    for i in demo_db:
-        if i["email_address"] == email_address:
-            is_users_email = i
-            break
-    if is_users_email is None:
-        raise HTTPException(status_code=401, detail="일치하는 아이디(이메일) 정보가 없습니다.")
-    # 2. 이메일은 확인되었고, 로그인 횟수 검증
-    attempts = login_attempts.get(email_address, 0)
-    if attempts >= MAX_ATTEMPTS:
-        raise HTTPException(status_code=429, detail="5회 이상 로그인에 실패하였습니다. 너무 많이 시도하였습니다.")
-    # 3. 아직 횟수 남았으면, 비밀번호 해시값 검증. 틀리면 횟수 1회 증가
-    if not verify_password(password, is_users_email["hashed_password"]):
-        login_attempts[email_address] = attempts + 1
-        raise HTTPException(status_code=401, detail="비밀번호가 맞지 않습니다. 다시 확인해 주세요.")
-    # 4. 비밀번호 해시값 검증 성공 시 초기화
-    login_attempts[email_address] = 0
-    # 5. 비밀번호 해시값 검증 완료, 액세스 토큰 & 리프레시 토큰 생성
-    access_token = create_access_token(
-        data={"sub":is_users_email["user_id"], "type" : "access"},
-        expires_delta=timedelta(minutes=int(os.getenv('ACCESS_TOKEN_EXPIRE_MINUTES')))
-    )
-    refresh_token = create_access_token(
-        data={"sub":is_users_email["user_id"], "type" : "refresh"},
-        expires_delta=timedelta(days=int(os.getenv('REFRESH_TOKEN_EXPIRE_DAYS')))
-    )
-    return {
-        "status": "success",
-        "message": "로그인에 성공하였습니다.",
-        "data": {
-            "user_id": is_users_email["user_id"],
-            "login_time": datetime.now(timezone.utc).strftime('%Y.%m.%d - %H:%M:%S'),
-            "token_type": "bearer",
-            "access_token": access_token,
-            "refresh_token": refresh_token,
-            "expires_in": int(os.getenv('ACCESS_TOKEN_EXPIRE_MINUTES')) * 60,
-            "issued_at": datetime.now(timezone.utc).strftime('%Y.%m.%d - %H:%M:%S')
-        }
-    }
-
-# ----- 세번째, 내 프로필 조회 메서드 구현
-@app.get("/users/my-page")
-async def get_users_my_page(user : dict = Depends(auth_user_with_token_decoding)):
-    return {
-        "status": "success",
-        "message": "액세스 토큰 및 사용자 인증이 완료되었습니다.",
-        "data": {
-            "user_id": user["user_id"],
-            "email_address" : user["email_address"],
-            "nickname": user["nickname"],
-            "profile_image_url": user["profile_image_url"],
-            "users_created_time": user["users_created_time"],
-            "users_modified_time": user["users_modified_time"]
-        }
-    }
-
-# ----- 네번째, 내 프로필 수정 메서드 구현
-@app.patch("/users/my-page")
-async def patch_users_my_page(
-        nickname : Annotated[str | None, Form()] = None,
-        password : Annotated[str | None, Form()] = None,
-        profile_image : Annotated[UploadFile | None, File()] = None,
-        user : dict = Depends(auth_user_with_token_decoding)
-):
-    # 최소 1개가 요청되었는지 확인
-    if not any([nickname, password, profile_image]):
-        raise HTTPException(status_code=400, detail="닉네임, 비밀번호, 프로필 이미지 중 최소 1개는 선택해야 합니다.")
-    # 1. 닉네임 변경이 요청된 경우 -- 일단 규칙 확인 후 중복 검사 (코드 리뷰에서 pythonic하게 변경)
-    if nickname:
-        validate_nickname(nickname) # 코드리뷰 반영 - 추가 검증 로직
-        if any(i["nickname"] == nickname and i["user_id"] != user["user_id"] for i in demo_db):
-            raise HTTPException(status_code=409, detail="해당 닉네임은 이미 존재합니다.")
-        user["nickname"] = nickname
-    # 2. 비밀번호 변경이 요청된 경우 -- 회원가입 로직과 동일함.
-    if password:
-        validate_password(password)
-        user["hashed_password"] = hash_password(password)
-    # 3. 이미지 변경이 요청된 경우 -- 회원가입 로직과 동일함.
-    if profile_image:
-        if profile_image.content_type not in ALLOWED_TYPES:
-            raise HTTPException(415, "지원하지 않는 이미지 파일 형식입니다. ")
-        #TODO : 파일 저장 및 이미지 크기 확인 로직
-        await validate_image_size(profile_image)
-        profile_image_url = f"/uploads/{profile_image.filename}"
-        user["profile_image_url"] = profile_image_url
-    # 4. 프로필 수정 시간 DB에 업데이트
-    user["users_modified_time"] = datetime.now(timezone.utc).strftime('%Y.%m.%d - %H:%M:%S')
-
-    return {
-        "status": "success",
-        "message": "프로필 수정이 완료되었습니다.",
-        "data": {
-            "user_id": user["user_id"],
-            "email_address" : user["email_address"],
-            "nickname": user["nickname"],
-            "profile_image_url": user["profile_image_url"],
-            "users_created_time": user["users_created_time"],
-            "users_modified_time": user["users_modified_time"]
-        }
-    }
-
-# ----- 다섯번째, 내 프로필 삭제 메서드 구현
-@app.delete("/users/my-page")
-async def delete_users_my_page(
-        user: dict = Depends(auth_user_with_token_decoding)
-):
-    # 데모 db에서 현재 사용자 제거
-    for i, j in enumerate(demo_db):
-        if j["user_id"] == user["user_id"]:
-            demo_db.remove(user)
-            break
-
-    return {
-        "status": "success",
-        "message": "유저가 탈퇴처리 되었습니다.",
-        "data": {
-            "user_id": user["user_id"]
-        }
-    }
-
-# ----- 여섯번째, 다른 사용자의 프로필 조회 메서드 구현
-@app.get("/users/{user_id}")
-async def get_users_user_id(user_id : str):
-    #로그인 인증 함수와 동일한 로직 적용
-    user = find_user_by_id(user_id)
-
-    return {
-        "status": "success",
-        "message": "요청한 회원 프로필 정보 조회에 성공했습니다.",
-        "data": {
-            "user_id": user["user_id"],
-            "email_address": user["email_address"],
-            "nickname": user["nickname"],
-            "profile_image_url": user["profile_image_url"],
-            "users_created_time": user["users_created_time"]
-        }
-    }
+from fastapi import FastAPIfrom routers import usersapp = FastAPI()# APIrouter를 활용하여 리팩토링 ------------------------ 2026.01.20# 라우터 등록app.include_router(users.router)# 정상 동작 여부 확인@app.get("/")def health_check():    return {        "status": "ok",        "message": "FastAPI 서버가 정상 동작중입니다. "    }
